Fix nav menu's `clickNested` clicking bug and callback.
This addresses a nested menu bug where the sub-navigation menu remains open (the
element remains visible) after a nested link is clicked. This resulted in odd
Selenium behavior with subsequent clicks and interactions. Adding a 'click' to
the page header effectively closes the sub-navigation menu.

This commit also implements a callback, even if one isn't provided via the
arguments.